+++ examples/nodes/call.py
@@ -46,10 +46,6 @@ Example 3:
             expr=Name(name='self')),
         args=[Name(name='x')],
         keywords=None)
-
-Type-checking:
-    The type of func must be a function type; the argument types are matched with the parameter types
-    of the function. The type of the Call expression itself is equal to the return type of the function.
 """
 
 # Example 1